refactor(command): added command help definitions
Added the help property to the Command class, inlined the constructor paramters
for the Command class and added documentation for the new parameters. Also refactored
the "SyncCommand" to implement the new constructor.

Also added the hint to the "execute" function, that subcommands will be searched. When a
sub command was found it will be called instead of the current command.