build(root): added the basic valid packagaya configuration! WOHO!
The configuration file uses the JSON schema which is defined in the
source code. The "apps" and the "libs" config entries are initialized
with their corresponding directories in the "packages" directory. Also
the "ts-adapter" will be used to process the source files and provide
different features for managing the source code. For now the
"sync-ts-paths" is enabled as feature flag.