feat(sync-ts-deps): added the "SyncTSDeps" feature flag
The "SyncTSDeps" feature flags reads the source code files and fetches all imports. The read imports
will than be compared to the actual defined dependencies in the package.json file. The user has to
answer some question (dependency type + version) when a dependency is missing. Dependencies which
are not used are removed from the configuration file.